+++ module-domain/domain-member/src/main/java/shop/jtoon/entity/Member.java
@@ -1,6 +1,7 @@
 package shop.jtoon.entity;
 
 import static java.util.Objects.*;
+import static shop.jtoon.util.RegExp.*;
 
 import java.time.LocalDateTime;
 
@@ -12,6 +13,7 @@ import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.Table;
+import jakarta.validation.constraints.Pattern;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
@@ -29,9 +31,11 @@ public class Member extends BaseTimeEntity {
 	@Column(name = "member_id")
 	private Long id;
 
+	@Pattern(regexp = EMAIL_PATTERN)
 	@Column(name = "email", nullable = false, unique = true, length = 40, updatable = false)
 	private String email;
 
+	@Pattern(regexp = PASSWORD_PATTERN)
 	@Column(name = "password", nullable = false)
 	private String password;
 
@@ -45,6 +49,7 @@ public class Member extends BaseTimeEntity {
 	@Column(name = "gender", nullable = false, updatable = false)
 	private Gender gender;
 
+	@Pattern(regexp = PHONE_PATTERN)
 	@Column(name = "phone", nullable = false, unique = true, length = 11)
 	private String phone;
