--- module-domain/domain-security/src/main/java/shop/jtoon/response/LoginRes.java
@@ -1,16 +1,0 @@
-package shop.jtoon.response;
-
-import lombok.Builder;
-
-@Builder
-public record LoginRes(
-	String accessToken,
-	String refreshToken
-) {
-	public static LoginRes of(String accessToken, String refreshToken) {
-		return LoginRes.builder()
-			.accessToken(accessToken)
-			.refreshToken(refreshToken)
-			.build();
-	}
-}
