+++ module-internal/core-web/src/main/java/shop/jtoon/config/SecurityConfig.java
@@ -16,14 +16,14 @@ import shop.jtoon.security.filter.JwtAuthenticationFilter;
 import shop.jtoon.security.handler.OAuth2FailureHandler;
 import shop.jtoon.security.handler.OAuth2SuccessHandler;
 import shop.jtoon.security.service.CustomOAuth2UserService;
-import shop.jtoon.security.service.JwtService;
+import shop.jtoon.security.service.JwtInternalService;
 
 @Configuration
 @RequiredArgsConstructor
 @EnableWebSecurity
 public class SecurityConfig {
 
-	private final JwtService jwtService;
+	private final JwtInternalService jwtInternalService;
 	private final HandlerExceptionResolver handlerExceptionResolver;
 	private final CustomOAuth2UserService customOAuth2UserService;
 	private final OAuth2SuccessHandler oAuth2SuccessHandler;
@@ -48,7 +48,7 @@ public class SecurityConfig {
 			.httpBasic(AbstractHttpConfigurer::disable)
 			.formLogin(AbstractHttpConfigurer::disable)
 			.sessionManagement(session -> session.sessionCreationPolicy(SessionCreationPolicy.STATELESS))
-			.addFilterBefore(new JwtAuthenticationFilter(handlerExceptionResolver, jwtService),
+			.addFilterBefore(new JwtAuthenticationFilter(handlerExceptionResolver, jwtInternalService),
 				UsernamePasswordAuthenticationFilter.class)
 			.oauth2Login(login -> login
 				.userInfoEndpoint(endpoint -> endpoint.userService(customOAuth2UserService))
