+++ module-internal/core-web/src/main/java/shop/jtoon/security/handler/OAuth2FailureHandler.java
@@ -1,5 +1,7 @@
 package shop.jtoon.security.handler;
 
+import java.io.IOException;
+
 import org.springframework.security.authentication.AuthenticationCredentialsNotFoundException;
 import org.springframework.security.authentication.BadCredentialsException;
 import org.springframework.security.authentication.InternalAuthenticationServiceException;
@@ -10,7 +12,6 @@ import org.springframework.stereotype.Component;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import java.io.IOException;
 import lombok.extern.slf4j.Slf4j;
 import shop.jtoon.type.ErrorStatus;
