+++ module-internal/core-web/src/main/java/shop/jtoon/security/handler/OAuth2SuccessHandler.java
@@ -2,6 +2,9 @@ package shop.jtoon.security.handler;
 
 import static shop.jtoon.util.SecurityConstant.*;
 
+import java.util.Map;
+import java.util.Objects;
+
 import org.springframework.security.core.Authentication;
 import org.springframework.security.oauth2.core.user.OAuth2User;
 import org.springframework.security.web.authentication.SimpleUrlAuthenticationSuccessHandler;
@@ -9,17 +12,17 @@ import org.springframework.stereotype.Component;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import java.util.Map;
-import java.util.Objects;
 import lombok.RequiredArgsConstructor;
-import shop.jtoon.security.service.JwtService;
+import shop.jtoon.security.service.JwtInternalService;
+import shop.jtoon.security.service.RefreshTokenService;
 import shop.jtoon.security.util.TokenCookie;
 
 @Component
 @RequiredArgsConstructor
 public class OAuth2SuccessHandler extends SimpleUrlAuthenticationSuccessHandler {
 
-	private final JwtService jwtService;
+	private final JwtInternalService jwtInternalService;
+	private final RefreshTokenService jwtService;
 
 	@Override
 	public void onAuthenticationSuccess(HttpServletRequest request, HttpServletResponse response,
@@ -27,8 +30,8 @@ public class OAuth2SuccessHandler extends SimpleUrlAuthenticationSuccessHandler 
 		OAuth2User oAuth2User = (OAuth2User)authentication.getPrincipal();
 		Map<String, Object> res = oAuth2User.getAttribute("response");
 		String email = (String)Objects.requireNonNull(res).get("email");
-		String accessToken = jwtService.generateAccessToken(email);
-		String refreshToken = jwtService.generateRefreshToken();
+		String accessToken = jwtInternalService.generateAccessToken(email);
+		String refreshToken = jwtInternalService.generateRefreshToken();
 		response.addCookie(TokenCookie.of(ACCESS_TOKEN_HEADER, accessToken));
 		response.addCookie(TokenCookie.of(REFRESH_TOKEN_HEADER, refreshToken));
 		jwtService.saveRefreshTokenDb(email, refreshToken);
