+++ module-internal/core-web/src/main/java/shop/jtoon/security/service/AuthenticationService.java
@@ -1,0 +1,7 @@
+package shop.jtoon.security.service;
+
+import org.springframework.security.core.Authentication;
+
+public interface AuthenticationService {
+	Authentication getAuthentication(String claimsEmail);
+}
