--- module-internal/core-web/src/main/java/shop/jtoon/security/service/JwtService.java
@@ -1,22 +1,0 @@
-package shop.jtoon.security.service;
-
-import org.springframework.security.core.Authentication;
-
-public interface JwtService {
-
-	String generateAccessToken(String email);
-
-	String reGenerateAccessToken(String refreshToken);
-
-	String generateRefreshToken();
-
-	boolean isTokenValid(String token);
-
-	Authentication getAuthentication(String token);
-
-	void verifyRefreshTokenDb(String refreshToken);
-
-	void updateRefreshTokenDb(String accessToken, String newRefreshToken);
-
-	void saveRefreshTokenDb(String email, String refreshToken);
-}
