+++ module-internal/core-web/src/main/java/shop/jtoon/security/service/RefreshTokenService.java
@@ -1,0 +1,12 @@
+package shop.jtoon.security.service;
+
+public interface RefreshTokenService {
+
+	String getRefreshTokenEmail(String refreshToken);
+
+	void verifyRefreshTokenDb(String refreshToken);
+
+	void updateRefreshTokenDb(String email, String newRefreshToken);
+
+	void saveRefreshTokenDb(String email, String refreshToken);
+}
